+++ src/app/components/Settings.js
@@ -37,6 +37,7 @@ import StyleIcon from "@mui/icons-material/Style";
 import PowerSettingsNewIcon from "@mui/icons-material/PowerSettingsNew";
 import AccountTreeIcon from "@mui/icons-material/AccountTree";
 import CloudQueueIcon from '@mui/icons-material/CloudQueue';
+import DeveloperModeIcon from '@mui/icons-material/DeveloperMode';
 
 import AccountsSettings from "./settings/AccountsSettings";
 import ProfileSettings from "./settings/ProfileSettings";
@@ -59,6 +60,8 @@ export default function Settings(props) {
   const location = useLocation();
   const dispatch = useDispatch();
   const theme = useTheme();
+
+  const isDeveloper = useSelector(state => state.app.isDeveloper);
 
   const SETTINGS = {
     PROFILE: {
@@ -126,6 +129,12 @@ export default function Settings(props) {
       url: "/settings/application/",
       subtitle: "Version, force refresh",
       icon: <SettingsApplications />,
+    },
+    DEVELOPER: {
+      title: "Dev tools",
+      url: "/settings/development/",
+      subtitle: "For testing and debugging",
+      icon: <DeveloperModeIcon />,
     },
     HELP: {
       title: "Help / Support",
@@ -254,7 +263,6 @@ export default function Settings(props) {
             </>
           )}
           </List>
-
           <List
             subheader={
               <ListSubheader disableSticky={true}>More settings</ListSubheader>
@@ -263,6 +271,16 @@ export default function Settings(props) {
             {drawListItem(SETTINGS.APP)}
             {drawListItem(SETTINGS.HELP)}
           </List>
+
+
+          { isDeveloper && <List
+            subheader={
+              <ListSubheader disableSticky={true}>Developement mode</ListSubheader>
+            }
+          >
+            { drawListItem(SETTINGS.DEVELOPER) }
+
+          </List>}
         </div>
 
         {page &&