+++ CHANGELOG.md
@@ -23,6 +23,7 @@ See for sample https://raw.githubusercontent.com/favoloso/conventional-changelog
 ## [1.2.0] - 2022-MM-DD
 ### 🛠 Improvements
 - Handle uncaught exception with a bug report view (#84)
+- List of categories now includes a no category value (#94)
 ### 🐛 Bug Fixes
 - Dashboard crash (#108)