+++ CHANGELOG.md
@@ -24,6 +24,7 @@ See for sample https://raw.githubusercontent.com/favoloso/conventional-changelog
 ### 🛠 Improvements
 - Handle uncaught exception with a bug report view (#84)
 - Implement a developer mode in app (#85)
+- Refactor the settings panel using react-router (#88)
 - List of categories includes a no category value (#94)
 ### 🐛 Bug Fixes
 - Fix Snackbar manager not working if no account (#87)