+++ CHANGELOG.md
@@ -33,6 +33,7 @@ See for sample https://raw.githubusercontent.com/favoloso/conventional-changelog
 - Fix Snackbar manager not working if no account (#87)
 - Dashboard crash (#108)
 - Add missing 'No category label' on mobile list of category (#111)
+- Fix npm install by updating dependencies (#113)
 
 ## [1.1.4] - 2022-02-27
 ### 🐛 Bug Fixes