+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 See for sample https://raw.githubusercontent.com/favoloso/conventional-changelog-emoji/master/CHANGELOG.md
 -->
-## [1.2.0] - 2022-MM-DD
+## [1.2.0] - 2022-07-06
 ### 🛠 Improvements
 - Handle uncaught exception with a bug report view (#84)
 - Implement a developer mode in app (#85)